refactoring db context, add identity user
